Upstreamed with https://github.com/j-hc/revanced-magisk-module
sort 'dev' patches versions

fix dl_uptodown

Update versionURL to include extraURL and versionID.

fix dl_uptodown

module: check 'unmount modules' and warn user

module: handle more install errors

module: unmount modules automatically

module: update

Update customize.sh

Update customize.sh

module: update

module: remove ksu_profile for now

Skip to the next mirror when 403 error occurs (#722)

When the Cloudflare protection triggers, the curl in __APKMIRROR_RESP__
receives a 403 error. However, the function doesn't exit imediately, as
a result, the last function exit code is the __APKMIRROR_CAT__ call that
can pass. In this situation, even if our mirror failed, the code skips
all other mirrors and exit imediately since the pkg_name is not
available.

The current modification makes sure that the function exits imediately
after the error, allowing us to move to the next mirror (uptodown)
instead of skipping the entire package build.

 get_apkmirror_resp() {
-       __APKMIRROR_RESP__=$(req "${1}" -)
+       __APKMIRROR_RESP__=$(req "${1}" -) || return 1
        __APKMIRROR_CAT__="${1##*/}"
 }

fix(#714): build fails when patch name contains a single quote (#715)

* fix(#714): by handle escaped and doubled single quotes in toml_get

* docs: add warning about single quotes in patch names

Allow multiple dpi for apkmirror (#718)

Introduce an option that allows multiple dpi versions to be
searched in order and make this option by default. The current
implementation allow this list to be changed from the config.toml
using a global variable like:
dpi = "nodpi anydpi 120-640dpi"

By default, only "nodpi" and "anydpi" are searched.

multipledpis: little fix

Fix uptodown download (#721)

This commit removes the limitations when there are multiple versions
available for a package. With this, all major applications should
be correctly identified and downloaded from uptodown.

This commit:
1. Increases the pages searched size from 5 to 20, due to the large
number of releases of each popular application (Facebook, Instagram,
Reddit, etc.). The performance impact for this increase is not
noticeable.
2. Check the file type to correctly identify if we are dealing with
a bundle or just an apk.
3. Adds an additional common architecture that was skipped due to our
strict filters.

Fix error message if the application is not found in the patches list (#732)

The current error message was always displaying an empty variable without
telling what could be the potential issue.
The proposed reimplementation tells exactly what the error is and shows
what applications are supported in the selected patches source. This
opens the path to support different patches sources that contain different
applcation lists, like Morphe.

Example output for Morphe patch source when trying to build Facebook:
[-] ABORT: No patch versions found for 'com.facebook.katana' in this patches source!
Available applications found: 'com.google.android.apps.youtube.music', 'com.google.android.youtube'

This change also removes the dead code around the op assignment since
in it's current implementation, that code would never execute. Additionally
the op value check was also done in the code sequence that was modified to
correctly display the error message.

Implement Morphe CLI and patches support (#734)

This commit restructures the code a bit to enable support of both
RVC and MPP (Morphe) client and patches.

The current implementation automatically determines the supported
patch type based on the selected cli application. This commit also
renames a number of variables / functions to better reflect the
state that they can reference both Revanced and Morphe.

Co-authored-by: j-hc <25510067+j-hc@users.noreply.github.com>

Error variable incorrectly refactored (#736)

After these 2 pull requests, 2 error message variables were not refactored:
https://github.com/j-hc/revanced-magisk-module/pull/734
https://github.com/j-hc/revanced-magisk-module/pull/732

Fix jq quoting

Check and warn about the asset number (#733)

The code despite allowing multiple assets to exist, it was designed
so that only 1 is expected for a release. This assumption held true,
until some releases like the Morphe's 1.1.0 release where they
incorrectly added 2 jars:
https://github.com/MorpheApp/morphe-cli/releases/tag/v1.1.0-dev.1

Since this is not expected (considering the fact that one of those
jars is named incorrectly), it's fair to assume that it's not the
intended way they want to do releases from now on and that we should
warn about these incorrect releases, even if we continue the execution
with the first found artifact.

Redirect GitHub error to stderr in epr() (#738)

The current implementation was sending the error to the stdout output.
This would making the situation when 2 tar artifacts were found fatal.
The error message was sent to stdout (while it was executed in Github)
and it was used as the return for the get_rv_prebuilts function.

This was not the intended behavior and this change fixes this bug.

introduce rvmm-zygisk-mount

update ksu_profile for KernelSU v3

add ksu_profile back

update patches-version for YouTube and YouTube Music to v4.0.0-dev.1